+++ adapters/Vpon/VponAdapter/GADVpadnNativeAdCustomEvent.h
@@ -1,0 +1,20 @@
+//
+//  GADVpadnNativeAdCustomEvent.h
+//  VponAdapter
+//
+//  Created by EricChien on 2018/10/24.
+//  Copyright © 2018 Vpon. All rights reserved.
+//
+
+@import GoogleMobileAds;
+@import VpadnSDKAdKit;
+
+@protocol GADCustomEventNativeAd;
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface GADVpadnNativeAdCustomEvent : NSObject
+
+@end
+
+NS_ASSUME_NONNULL_END