+++ adapters/Vpon/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Vpon Mediation Adapter for Google Mobile Ads SDK for iOS
 
-## Version x.y.z.a
-- Under development
+#### [Version 2.0.6.0](https://m.vpon.com/sdk/ios/VpadnSDKiOS-5.3.2.zip)
+- Updated the adapter to use the `.xcframework` format.
+- Upload to googleads-mobile-ios-mediation.
+
+Built and tested with:
+- Google Mobile Ads SDK version 8.13.0.
+- Vpon SDK version 5.3.2.